moved spirv tests into front/spv/mod.rs
