+++ naga/src/front/atomic_upgrade.rs
@@ -212,82 +212,3 @@ impl Module {
         Ok(())
     }
 }
-
-#[cfg(test)]
-mod test {
-    fn atomic_test(bytes: &[u8]) {
-        let _ = env_logger::builder().is_test(true).try_init();
-        let m = crate::front::spv::parse_u8_slice(bytes, &Default::default()).unwrap();
-
-        let mut wgsl = String::new();
-        let mut should_panic = false;
-
-        for vflags in [
-            crate::valid::ValidationFlags::all(),
-            crate::valid::ValidationFlags::empty(),
-        ] {
-            let mut validator = crate::valid::Validator::new(vflags, Default::default());
-            match validator.validate(&m) {
-                Err(e) => {
-                    log::error!("SPIR-V validation {}", e.emit_to_string(""));
-                    should_panic = true;
-                }
-                Ok(i) => {
-                    wgsl = crate::back::wgsl::write_string(
-                        &m,
-                        &i,
-                        crate::back::wgsl::WriterFlags::empty(),
-                    )
-                    .unwrap();
-                    log::info!("wgsl-out:\n{wgsl}");
-                    break;
-                }
-            };
-        }
-
-        if should_panic {
-            panic!("validation error");
-        }
-
-        let m = match crate::front::wgsl::parse_str(&wgsl) {
-            Ok(m) => m,
-            Err(e) => {
-                log::error!("round trip WGSL validation {}", e.emit_to_string(&wgsl));
-                panic!("invalid module");
-            }
-        };
-        let mut validator =
-            crate::valid::Validator::new(crate::valid::ValidationFlags::all(), Default::default());
-        if let Err(e) = validator.validate(&m) {
-            log::error!("{}", e.emit_to_string(&wgsl));
-            panic!("invalid generated wgsl");
-        }
-    }
-
-    #[test]
-    fn atomic_i_inc() {
-        atomic_test(include_bytes!("../../tests/in/spv/atomic_i_increment.spv"));
-    }
-
-    #[test]
-    fn atomic_load_and_store() {
-        atomic_test(include_bytes!(
-            "../../tests/in/spv/atomic_load_and_store.spv"
-        ));
-    }
-
-    #[test]
-    fn atomic_exchange() {
-        atomic_test(include_bytes!("../../tests/in/spv/atomic_exchange.spv"));
-    }
-
-    #[test]
-    fn atomic_i_decrement() {
-        atomic_test(include_bytes!("../../tests/in/spv/atomic_i_decrement.spv"));
-    }
-
-    #[test]
-    fn atomic_i_add_and_sub() {
-        atomic_test(include_bytes!("../../tests/in/spv/atomic_i_add_sub.spv"));
-    }
-}
